refactor(dist): move generated backup api models to a subpackage
This helps to keep the classes related to backup and cluster management in different packages
