+++ qa/integration-tests/src/test/java/io/camunda/zeebe/it/backup/S3RestoreAcceptanceIT.java
@@ -15,6 +15,9 @@ import io.camunda.zeebe.backup.s3.S3BackupConfig.Builder;
 import io.camunda.zeebe.backup.s3.S3BackupStore;
 import io.camunda.zeebe.broker.system.configuration.BrokerCfg;
 import io.camunda.zeebe.broker.system.configuration.backup.BackupStoreCfg.BackupStoreType;
+import io.camunda.zeebe.management.backups.BackupInfo;
+import io.camunda.zeebe.management.backups.StateCode;
+import io.camunda.zeebe.management.backups.TakeBackupResponse;
 import io.camunda.zeebe.qa.util.actuator.BackupActuator;
 import io.camunda.zeebe.qa.util.cluster.TestRestoreApp;
 import io.camunda.zeebe.qa.util.cluster.TestStandaloneBroker;
@@ -22,9 +25,6 @@ import io.camunda.zeebe.qa.util.junit.ZeebeIntegration;
 import io.camunda.zeebe.qa.util.junit.ZeebeIntegration.TestZeebe;
 import io.camunda.zeebe.qa.util.testcontainers.MinioContainer;
 import io.camunda.zeebe.restore.BackupNotFoundException;
-import io.camunda.zeebe.shared.management.openapi.models.BackupInfo;
-import io.camunda.zeebe.shared.management.openapi.models.StateCode;
-import io.camunda.zeebe.shared.management.openapi.models.TakeBackupResponse;
 import java.time.Duration;
 import org.apache.commons.lang3.RandomStringUtils;
 import org.awaitility.Awaitility;
